Removed unused icon and refactor in test files
