+++ android/app/src/androidTest/java/com/votlin/android/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package org.jetbrains.kotlinconf
+package com.votlin.android
 
 import android.support.test.InstrumentationRegistry
 import android.support.test.runner.AndroidJUnit4
@@ -19,6 +19,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getTargetContext()
-        assertEquals("org.jetbrains.kotlinconf", appContext.packageName)
+        assertEquals("com.votlin.android", appContext.packageName)
     }
 }
